Move the complex shrinker case to a specific file
